TC#10: Refactored for hooks structure
