+++ frontend/src/features/news/OneNews.tsx
@@ -1,33 +1,17 @@
-import React, { useEffect } from 'react';
-import { useParams } from 'react-router-dom';
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import useEmblaCarousel from 'embla-carousel-react';
-import { fetchNewsByLimit, fetchOneNews } from '@/features/news/newsThunks';
-import { selectNews, selectOneNews } from '@/features/news/newsSlice';
+import React from 'react';
 import { API_URl } from '@/consts';
-import { NextButton, PrevButton, usePrevNextButtons } from '@/features/news/components/CarouselButtons';
-import NewsCard from '@/features/news/components/NewsCard';
-import './embla.css';
+import { useOneNews } from '@/features/news/hooks/useOneNews';
+import { Layout } from '@/components/Layout';
+import { NextButton, PrevButton } from '@/features/news/components/CustomCarousel/CarouselButtons';
+import { NewsCard } from '@/features/news/components/NewsCard/NewsCard';
+import './components/CustomCarousel/embla.css';
 
 export const OneNews: React.FC = () => {
-  const [emblaRef, emblaApi] = useEmblaCarousel({ dragFree: true, loop: true });
-  const { prevBtnDisabled, nextBtnDisabled, onPrevButtonClick, onNextButtonClick } = usePrevNextButtons(emblaApi);
-
-  const { id } = useParams();
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    if (id) {
-      dispatch(fetchOneNews(id));
-      dispatch(fetchNewsByLimit(4));
-    }
-  }, [dispatch, id]);
-
-  const oneNews = useAppSelector(selectOneNews);
-  const news = useAppSelector(selectNews);
+  const { emblaRef, oneNews, news, prevBtnDisabled, nextBtnDisabled, onPrevButtonClick, onNextButtonClick } =
+    useOneNews();
 
   return (
-    <main className='py-16'>
+    <Layout className='py-16'>
       <div className='text-center mb-5 sm:mb-12'>
         <h1 className='font-semibold leading-10 text-2xl sm:text-4xl md:text-5xl sm:mb-6'>{oneNews?.subtitle}</h1>
         <h2 className='text-cr-gray-500 text-[20px] sm:text-2xl font-medium uppercase'>{oneNews?.title}</h2>
@@ -50,6 +34,7 @@ export const OneNews: React.FC = () => {
           </div>
         </div>
       </section>
+
       <section className='mb-5'>
         <div className='text-lg sm:text-xl leading-6 sm:leading-8 mb-5'>{oneNews?.content}</div>
       </section>
@@ -61,6 +46,6 @@ export const OneNews: React.FC = () => {
           ))}
         </div>
       </section>
-    </main>
+    </Layout>
   );
 };
